# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImpactZoneCollection": {
        "properties": {
          "features": {
            "items": {
              "$ref": "#/components/schemas/ImpactZoneFeature"
            },
            "title": "Features",
            "type": "array"
          },
          "type": {
            "const": "FeatureCollection",
            "default": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "ImpactZoneCollection",
        "type": "object"
      },
      "ImpactZoneFeature": {
        "properties": {
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "properties": {
            "$ref": "#/components/schemas/ImpactZoneProps"
          },
          "type": {
            "const": "Feature",
            "default": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "properties"
        ],
        "title": "ImpactZoneFeature",
        "type": "object"
      },
      "ImpactZoneProps": {
        "description": "Egenskaper för en serverberäknad påverkanszon (buffrad projektgeometri).",
        "properties": {
          "impact_radius_m": {
            "title": "Impact Radius M",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "project_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectType"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "project_id",
          "name",
          "impact_radius_m"
        ],
        "title": "ImpactZoneProps",
        "type": "object"
      },
      "InfrastructureProjectCollection": {
        "properties": {
          "features": {
            "items": {
              "$ref": "#/components/schemas/InfrastructureProjectFeature"
            },
            "title": "Features",
            "type": "array"
          },
          "numberMatched": {
            "description": "Totalt antal träffar för frågan",
            "title": "Numbermatched",
            "type": "integer"
          },
          "numberReturned": {
            "description": "Antal features i detta svar",
            "title": "Numberreturned",
            "type": "integer"
          },
          "type": {
            "const": "FeatureCollection",
            "default": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "numberMatched",
          "numberReturned"
        ],
        "title": "InfrastructureProjectCollection",
        "type": "object"
      },
      "InfrastructureProjectCreate": {
        "properties": {
          "budget_sek": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Sek"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "GeoJSON-geometri i WGS84",
            "title": "Geometry"
          },
          "impact_radius_m": {
            "default": 1000.0,
            "exclusiveMinimum": 0.0,
            "maximum": 50000.0,
            "title": "Impact Radius M",
            "type": "number"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "project_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectType"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "default": "manual",
            "title": "Source",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name"
        ],
        "title": "InfrastructureProjectCreate",
        "type": "object"
      },
      "InfrastructureProjectFeature": {
        "properties": {
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "properties": {
            "$ref": "#/components/schemas/InfrastructureProjectProps"
          },
          "type": {
            "const": "Feature",
            "default": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "properties"
        ],
        "title": "InfrastructureProjectFeature",
        "type": "object"
      },
      "InfrastructureProjectProps": {
        "properties": {
          "budget_sek": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Sek"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "impact_radius_m": {
            "default": 1000.0,
            "title": "Impact Radius M",
            "type": "number"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectType"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "default": "manual",
            "title": "Source",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "InfrastructureProjectProps",
        "type": "object"
      },
...
      "ProjectStatus": {
        "enum": [
          "planerad",
          "pågående",
          "avslutad"
        ],
        "title": "ProjectStatus",
        "type": "string"
      },
      "ProjectType": {
        "enum": [
          "väg",
          "järnväg",
          "kollektivtrafik",
          "bro",
          "tunnel",
          "cykelväg",
          "övrigt"
        ],
        "title": "ProjectType",
        "type": "string"
      },
...
      "SyncResult": {
        "description": "Resultatet av en synkronisering mot en extern datakälla.",
        "properties": {
          "fetched": {
            "description": "Antal objekt som datakällan levererade",
            "title": "Fetched",
            "type": "integer"
          },
          "skipped": {
            "description": "Antal objekt som hoppades över (t.ex. utan externt id)",
            "title": "Skipped",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "upserted": {
            "description": "Antal objekt som skapades eller uppdaterades",
            "title": "Upserted",
            "type": "integer"
          }
        },
        "required": [
          "source",
          "fetched",
          "upserted",
          "skipped"
        ],
        "title": "SyncResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API för visualisering av svenska fastigheter och infrastrukturprojekt. Geodata lagras i PostGIS och serveras som GeoJSON.",
    "title": "Fastighetsvisualiserare API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/infrastructure/impact-zones": {
      "get": {
        "description": "Påverkanszoner: projektgeometrier buffrade med sin påverkansradie i meter.\n\nBeräknas i PostGIS över geography, vilket ger korrekta zoner för\npunkter, linjer och ytor oavsett latitud.",
        "operationId": "impact_zones_api_v1_infrastructure_impact_zones_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ProjectStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "project_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ProjectType"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Type"
            }
          },
          {
            "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactZoneCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Impact Zones",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/infrastructure/projects": {
      "get": {
        "description": "Infrastrukturprojekt som GeoJSON FeatureCollection, med filter och paginering.",
        "operationId": "list_projects_api_v1_infrastructure_projects_get",
        "parameters": [
          {
            "description": "Filtrera på status (kan upprepas)",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ProjectStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrera på status (kan upprepas)",
              "title": "Status"
            }
          },
          {
            "description": "Filtrera på projekttyp (kan upprepas)",
            "in": "query",
            "name": "project_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ProjectType"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrera på projekttyp (kan upprepas)",
              "title": "Project Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfrastructureProjectCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "infrastructure"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_infrastructure_projects_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfrastructureProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfrastructureProjectFeature"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/infrastructure/projects/{project_id}": {
      "get": {
        "operationId": "get_project_api_v1_infrastructure_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfrastructureProjectFeature"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "infrastructure"
        ]
      }
    },
...
    "/api/v1/infrastructure/sync/{source_name}": {
      "post": {
        "description": "Synkronisera en registrerad datakälla till databasen (upsert).",
        "operationId": "sync_source_api_v1_infrastructure_sync__source_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "source_name",
            "required": true,
            "schema": {
              "title": "Source Name",
              "type": "string"
            }
          },
          {
            "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
              "title": "Bbox"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Source",
        "tags": [
          "infrastructure"
        ]
      }
    },
...
  }
}